refactor: remove old processor, start fetching data from contract
