+++ .env
@@ -122,34 +122,11 @@ STACKS_NODE_TYPE=L1
 # Override the default file path for the proxy cache control file
 # STACKS_API_PROXY_CACHE_CONTROL_FILE=/path/to/.proxy-cache-control.json
 
-# Enable token metadata processing. Disabled by default.
-# STACKS_API_ENABLE_FT_METADATA=1
-# STACKS_API_ENABLE_NFT_METADATA=1
-
-# If token metadata processing is enabled, this variable determines how the API reacts to metadata processing failures.
-# When strict mode is enabled, any failures caused by recoverable errors will be retried indefinitely. Otherwise,
-# the API will give up after `STACKS_API_TOKEN_METADATA_MAX_RETRIES` is reached for that smart contract.
-# STACKS_API_TOKEN_METADATA_STRICT_MODE=1
-
-# Maximum number of times we'll try processing FT/NFT metadata for a specific smart contract if we've failed
-# because of a recoverable error.
-# Only used if `STACKS_API_TOKEN_METADATA_STRICT_MODE` is disabled.
-# STACKS_API_TOKEN_METADATA_MAX_RETRIES=5
-
 # Controls the token metadata error handling mode. The possible values are:
 # * `warning`: If required metadata is not found, the API will issue a warning and not display data for that token.
 # * `error`: If required metadata is not found, the API will throw an error.
 # If not specified or any other value is provided, the mode will be set to `warning`.
 # STACKS_API_TOKEN_METADATA_ERROR_MODE=warning
-
-# Configure a script to handle image URLs during token metadata processing.
-# This example script uses the `imgix.net` service to create CDN URLs.
-# Must be an executable script that accepts the URL as the first program argument
-# and outputs a result URL to stdout.
-# STACKS_API_IMAGE_CACHE_PROCESSOR=./config/token-metadata-image-cache-imgix.js
-# Env vars needed for the above sample `imgix` script:
-# IMGIX_DOMAIN=https://<your domain>.imgix.net
-# IMGIX_TOKEN=<your token>
 
 # Web Socket ping interval to determine client availability, in seconds.
 # STACKS_API_WS_PING_INTERVAL=5
