+++ src/api/init.ts
@@ -31,7 +31,7 @@ import * as expressListEndpoints from 'express-list-endpoints';
 import { createMiddleware as createPrometheusMiddleware } from '@promster/express';
 import { createMicroblockRouter } from './routes/microblock';
 import { createStatusRouter } from './routes/status';
-import { createTokenRouter } from './routes/tokens/tokens';
+import { createTokenRouter } from './routes/tokens';
 import { createFeeRateRouter } from './routes/fee-rate';
 import { setResponseNonCacheable } from './controllers/cache-controller';
