+++ src/datastore/pg-write-store.ts
@@ -22,9 +22,6 @@ import {
   DataStoreMicroblockUpdateData,
   DbMicroblock,
   DataStoreTxEventData,
-  DbNonFungibleTokenMetadata,
-  DbFungibleTokenMetadata,
-  DbTokenMetadataQueueEntry,
   DbFaucetRequest,
   MinerRewardInsertValues,
   BlockInsertValues,
@@ -42,12 +39,9 @@ import {
   TxInsertValues,
   MempoolTxInsertValues,
   MempoolTxQueryResult,
-  TokenMetadataQueueEntryInsertValues,
   SmartContractInsertValues,
   BnsNameInsertValues,
   BnsNamespaceInsertValues,
-  FtMetadataInsertValues,
-  NftMetadataInsertValues,
   FaucetRequestInsertValues,
   MicroblockInsertValues,
   TxQueryResult,
@@ -66,7 +60,6 @@ import {
   IndexesState,
   NftCustodyInsertValues,
 } from './common';
-import { ClarityAbi } from '@stacks/transactions';
 import {
   BLOCK_COLUMNS,
   convertTxQueryResultToDbMempoolTx,
@@ -82,7 +75,6 @@ import {
 } from './helpers';
 import { PgNotifier } from './pg-notifier';
 import { MIGRATIONS_DIR, PgStore } from './pg-store';
-import { isProcessableTokenMetadata } from '../token-metadata/helpers';
 import * as zoneFileParser from 'zone-file';
 import { parseResolver, parseZoneFileTxt } from '../event-stream/bns/bns-helpers';
 import { Pox2EventName } from '../pox-helpers';
@@ -200,7 +192,6 @@ export class PgWriteStore extends PgStore {
   }
 
   async update(data: DataStoreBlockUpdateData): Promise<void> {
-    const tokenMetadataQueueEntries: DbTokenMetadataQueueEntry[] = [];
     let garbageCollectedMempoolTxs: string[] = [];
     let batchedTxData: DataStoreTxEventData[] = [];
     const deployedSmartContracts: DbSmartContract[] = [];
@@ -418,34 +409,6 @@ export class PgWriteStore extends PgStore {
           logger.debug(`Garbage collected ${mempoolGarbageResults.deletedTxs.length} mempool txs`);
         }
         garbageCollectedMempoolTxs = mempoolGarbageResults.deletedTxs;
-
-        const tokenContractDeployments = data.txs
-          .filter(
-            entry =>
-              entry.tx.type_id === DbTxTypeId.SmartContract ||
-              entry.tx.type_id === DbTxTypeId.VersionedSmartContract
-          )
-          .filter(entry => entry.tx.status === DbTxStatus.Success)
-          .filter(entry => entry.smartContracts[0].abi && entry.smartContracts[0].abi !== 'null')
-          .map(entry => {
-            const smartContract = entry.smartContracts[0];
-            const contractAbi: ClarityAbi = JSON.parse(smartContract.abi as string);
-            const queueEntry: DbTokenMetadataQueueEntry = {
-              queueId: -1,
-              txId: entry.tx.tx_id,
-              contractId: smartContract.contract_id,
-              contractAbi: contractAbi,
-              blockHeight: entry.tx.block_height,
-              processed: false,
-              retry_count: 0,
-            };
-            return queueEntry;
-          })
-          .filter(entry => isProcessableTokenMetadata(entry.contractAbi));
-        for (const pendingQueueEntry of tokenContractDeployments) {
-          const queueEntry = await this.updateTokenMetadataQueue(sql, pendingQueueEntry);
-          tokenMetadataQueueEntries.push(queueEntry);
-        }
       }
 
       if (!this.isEventReplay) {
@@ -489,9 +452,6 @@ export class PgWriteStore extends PgStore {
           txId: nftEvent.tx_id,
           eventIndex: nftEvent.event_index,
         });
-      }
-      for (const tokenMetadataQueueEntry of tokenMetadataQueueEntries) {
-        await this.notifier.sendTokenMetadata({ queueId: tokenMetadataQueueEntry.queueId });
       }
     }
   }
@@ -1841,27 +1801,27 @@ export class PgWriteStore extends PgStore {
     }
   }
 
-  async updateTokenMetadataQueue(
-    sql: PgSqlClient,
-    entry: DbTokenMetadataQueueEntry
-  ): Promise<DbTokenMetadataQueueEntry> {
-    const values: TokenMetadataQueueEntryInsertValues = {
-      tx_id: entry.txId,
-      contract_id: entry.contractId,
-      contract_abi: JSON.stringify(entry.contractAbi),
-      block_height: entry.blockHeight,
-      processed: false,
-    };
-    const queryResult = await sql<{ queue_id: number }[]>`
-      INSERT INTO token_metadata_queue ${sql(values)}
-      RETURNING queue_id
-    `;
-    const result: DbTokenMetadataQueueEntry = {
-      ...entry,
-      queueId: queryResult[0].queue_id,
-    };
-    return result;
-  }
+  // async updateTokenMetadataQueue(
+  //   sql: PgSqlClient,
+  //   entry: DbTokenMetadataQueueEntry
+  // ): Promise<DbTokenMetadataQueueEntry> {
+  //   const values: TokenMetadataQueueEntryInsertValues = {
+  //     tx_id: entry.txId,
+  //     contract_id: entry.contractId,
+  //     contract_abi: JSON.stringify(entry.contractAbi),
+  //     block_height: entry.blockHeight,
+  //     processed: false,
+  //   };
+  //   const queryResult = await sql<{ queue_id: number }[]>`
+  //     INSERT INTO token_metadata_queue ${sql(values)}
+  //     RETURNING queue_id
+  //   `;
+  //   const result: DbTokenMetadataQueueEntry = {
+  //     ...entry,
+  //     queueId: queryResult[0].queue_id,
+  //   };
+  //   return result;
+  // }
 
   async updateSmartContract(sql: PgSqlClient, tx: DbTx, smartContract: DbSmartContract) {
     const values: SmartContractInsertValues = {
@@ -2061,86 +2021,86 @@ export class PgWriteStore extends PgStore {
     `;
   }
 
-  async updateFtMetadata(ftMetadata: DbFungibleTokenMetadata, dbQueueId: number): Promise<number> {
-    const length = await this.sqlWriteTransaction(async sql => {
-      const values: FtMetadataInsertValues = {
-        token_uri: ftMetadata.token_uri,
-        name: ftMetadata.name,
-        description: ftMetadata.description,
-        image_uri: ftMetadata.image_uri,
-        image_canonical_uri: ftMetadata.image_canonical_uri,
-        contract_id: ftMetadata.contract_id,
-        symbol: ftMetadata.symbol,
-        decimals: ftMetadata.decimals,
-        tx_id: ftMetadata.tx_id,
-        sender_address: ftMetadata.sender_address,
-      };
-      const result = await sql`
-        INSERT INTO ft_metadata ${sql(values)}
-        ON CONFLICT (contract_id)
-        DO
-          UPDATE SET ${sql(values)}
-      `;
-      await sql`
-        UPDATE token_metadata_queue
-        SET processed = true
-        WHERE queue_id = ${dbQueueId}
-      `;
-      return result.count;
-    });
-    await this.notifier?.sendTokens({ contractID: ftMetadata.contract_id });
-    return length;
-  }
+  // async updateFtMetadata(ftMetadata: DbFungibleTokenMetadata, dbQueueId: number): Promise<number> {
+  //   const length = await this.sqlWriteTransaction(async sql => {
+  //     const values: FtMetadataInsertValues = {
+  //       token_uri: ftMetadata.token_uri,
+  //       name: ftMetadata.name,
+  //       description: ftMetadata.description,
+  //       image_uri: ftMetadata.image_uri,
+  //       image_canonical_uri: ftMetadata.image_canonical_uri,
+  //       contract_id: ftMetadata.contract_id,
+  //       symbol: ftMetadata.symbol,
+  //       decimals: ftMetadata.decimals,
+  //       tx_id: ftMetadata.tx_id,
+  //       sender_address: ftMetadata.sender_address,
+  //     };
+  //     const result = await sql`
+  //       INSERT INTO ft_metadata ${sql(values)}
+  //       ON CONFLICT (contract_id)
+  //       DO
+  //         UPDATE SET ${sql(values)}
+  //     `;
+  //     await sql`
+  //       UPDATE token_metadata_queue
+  //       SET processed = true
+  //       WHERE queue_id = ${dbQueueId}
+  //     `;
+  //     return result.count;
+  //   });
+  //   await this.notifier?.sendTokens({ contractID: ftMetadata.contract_id });
+  //   return length;
+  // }
 
-  async updateNFtMetadata(
-    nftMetadata: DbNonFungibleTokenMetadata,
-    dbQueueId: number
-  ): Promise<number> {
-    const length = await this.sqlWriteTransaction(async sql => {
-      const values: NftMetadataInsertValues = {
-        token_uri: nftMetadata.token_uri,
-        name: nftMetadata.name,
-        description: nftMetadata.description,
-        image_uri: nftMetadata.image_uri,
-        image_canonical_uri: nftMetadata.image_canonical_uri,
-        contract_id: nftMetadata.contract_id,
-        tx_id: nftMetadata.tx_id,
-        sender_address: nftMetadata.sender_address,
-      };
-      const result = await sql`
-        INSERT INTO nft_metadata ${sql(values)}
-        ON CONFLICT (contract_id)
-        DO
-          UPDATE SET ${sql(values)}
-      `;
-      await sql`
-        UPDATE token_metadata_queue
-        SET processed = true
-        WHERE queue_id = ${dbQueueId}
-      `;
-      return result.count;
-    });
-    await this.notifier?.sendTokens({ contractID: nftMetadata.contract_id });
-    return length;
-  }
+  // async updateNFtMetadata(
+  //   nftMetadata: DbNonFungibleTokenMetadata,
+  //   dbQueueId: number
+  // ): Promise<number> {
+  //   const length = await this.sqlWriteTransaction(async sql => {
+  //     const values: NftMetadataInsertValues = {
+  //       token_uri: nftMetadata.token_uri,
+  //       name: nftMetadata.name,
+  //       description: nftMetadata.description,
+  //       image_uri: nftMetadata.image_uri,
+  //       image_canonical_uri: nftMetadata.image_canonical_uri,
+  //       contract_id: nftMetadata.contract_id,
+  //       tx_id: nftMetadata.tx_id,
+  //       sender_address: nftMetadata.sender_address,
+  //     };
+  //     const result = await sql`
+  //       INSERT INTO nft_metadata ${sql(values)}
+  //       ON CONFLICT (contract_id)
+  //       DO
+  //         UPDATE SET ${sql(values)}
+  //     `;
+  //     await sql`
+  //       UPDATE token_metadata_queue
+  //       SET processed = true
+  //       WHERE queue_id = ${dbQueueId}
+  //     `;
+  //     return result.count;
+  //   });
+  //   await this.notifier?.sendTokens({ contractID: nftMetadata.contract_id });
+  //   return length;
+  // }
 
-  async updateProcessedTokenMetadataQueueEntry(queueId: number): Promise<void> {
-    await this.sql`
-      UPDATE token_metadata_queue
-      SET processed = true
-      WHERE queue_id = ${queueId}
-    `;
-  }
+  // async updateProcessedTokenMetadataQueueEntry(queueId: number): Promise<void> {
+  //   await this.sql`
+  //     UPDATE token_metadata_queue
+  //     SET processed = true
+  //     WHERE queue_id = ${queueId}
+  //   `;
+  // }
 
-  async increaseTokenMetadataQueueEntryRetryCount(queueId: number): Promise<number> {
-    const result = await this.sql<{ retry_count: number }[]>`
-      UPDATE token_metadata_queue
-      SET retry_count = retry_count + 1
-      WHERE queue_id = ${queueId}
-      RETURNING retry_count
-    `;
-    return result[0].retry_count;
-  }
+  // async increaseTokenMetadataQueueEntryRetryCount(queueId: number): Promise<number> {
+  //   const result = await this.sql<{ retry_count: number }[]>`
+  //     UPDATE token_metadata_queue
+  //     SET retry_count = retry_count + 1
+  //     WHERE queue_id = ${queueId}
+  //     RETURNING retry_count
+  //   `;
+  //   return result[0].retry_count;
+  // }
 
   async updateBatchTokenOfferingLocked(sql: PgSqlClient, lockedInfos: DbTokenOfferingLocked[]) {
     try {
