+++ src/rosetta/rosetta-helpers.ts
@@ -25,26 +25,23 @@ import { StacksMainnet, StacksTestnet } from '@stacks/network';
 import { ec as EC } from 'elliptic';
 import * as btc from 'bitcoinjs-lib';
 import {
-  getAssetEventTypeString,
-  getEventTypeString,
   getTxFromDataStore,
   getTxStatus,
   getTxTypeString,
   parseContractCallMetadata,
-} from './api/controllers/db-controller';
+} from '../api/controllers/db-controller';
 import {
   PoxContractIdentifier,
   RosettaConstants,
   RosettaNetworks,
   RosettaOperationType,
-} from './api/rosetta-constants';
+} from '../api/rosetta-constants';
 import {
   BaseTx,
   DbAssetEventTypeId,
   DbEvent,
   DbEventTypeId,
   DbFtEvent,
-  DbFungibleTokenMetadata,
   DbMempoolTx,
   DbMinerReward,
   DbStxEvent,
@@ -53,12 +50,11 @@ import {
   DbTxStatus,
   DbTxTypeId,
   StxUnlockEvent,
-} from './datastore/common';
-import { getTxSenderAddress, getTxSponsorAddress } from './event-stream/reader';
-import { unwrapOptional, getSendManyContract } from './helpers';
+} from '../datastore/common';
+import { getTxSenderAddress, getTxSponsorAddress } from '../event-stream/reader';
+import { unwrapOptional, getSendManyContract } from '../helpers';
 
-import { getCoreNodeEndpoint } from './core-rpc/client';
-import { TokenMetadataErrorMode } from './token-metadata/tokens-contract-handler';
+import { getCoreNodeEndpoint } from '../core-rpc/client';
 import {
   ClarityTypeID,
   decodeClarityValue,
@@ -76,12 +72,12 @@ import {
   ClarityValue,
   ClarityValueList,
 } from 'stacks-encoding-native-js';
-import { PgStore } from './datastore/pg-store';
-import { isFtMetadataEnabled, tokenMetadataErrorMode } from './token-metadata/helpers';
+import { PgStore } from '../datastore/pg-store';
 import { poxAddressToBtcAddress } from '@stacks/stacking';
 import { parseRecoverableSignatureVrs } from '@stacks/common';
-import { logger } from './logger';
+import { logger } from '../logger';
 import { hexToBuffer } from '@hirosystems/api-toolkit';
+import { RosettaFtMetadata, RosettaFtMetadataClient } from './rosetta-ft-metadata-client';
 
 enum CoinAction {
   CoinSpent = 'coin_spent',
@@ -243,6 +239,7 @@ async function processEvents(
   // match them by index.
   const sendManyMemos = decodeSendManyContractCallMemos(baseTx, chainID);
   let sendManyStxTransferEventIndex = 0;
+  const metadataClient = new RosettaFtMetadataClient(chainID);
 
   for (const event of events) {
     const txEventType = event.event_type;
@@ -298,7 +295,7 @@ async function processEvents(
       case DbEventTypeId.NonFungibleTokenAsset:
         break;
       case DbEventTypeId.FungibleTokenAsset:
-        const ftMetadata = await getValidatedFtMetadata(db, event.asset_identifier);
+        const ftMetadata = await metadataClient.getFtMetadata(event.asset_identifier);
         if (!ftMetadata) {
           break;
         }
@@ -435,7 +432,7 @@ function makeBurnOperation(tx: DbStxEvent, baseTx: BaseTx, index: number): Roset
 
 function makeFtBurnOperation(
   ftEvent: DbFtEvent,
-  ftMetadata: FungibleTokenMetadata,
+  ftMetadata: RosettaFtMetadata,
   baseTx: BaseTx,
   index: number
 ): RosettaOperation {
@@ -482,7 +479,7 @@ function makeMintOperation(tx: DbStxEvent, baseTx: BaseTx, index: number): Roset
 
 function makeFtMintOperation(
   ftEvent: DbFtEvent,
-  ftMetadata: FungibleTokenMetadata,
+  ftMetadata: RosettaFtMetadata,
   baseTx: BaseTx,
   index: number
 ): RosettaOperation {
@@ -548,7 +545,7 @@ function makeSenderOperation(
 
 function makeFtSenderOperation(
   ftEvent: DbFtEvent,
-  ftMetadata: FungibleTokenMetadata,
+  ftMetadata: RosettaFtMetadata,
   tx: BaseTx,
   index: number
 ): RosettaOperation {
@@ -621,7 +618,7 @@ function makeReceiverOperation(
 
 function makeFtReceiverOperation(
   ftEvent: DbFtEvent,
-  ftMetadata: FungibleTokenMetadata,
+  ftMetadata: RosettaFtMetadata,
   tx: BaseTx,
   index: number
 ): RosettaOperation {
@@ -1146,28 +1143,6 @@ export function rawTxToBaseTx(raw_tx: string): BaseTx {
   }
 
   return dbTx;
-}
-
-export async function getValidatedFtMetadata(
-  db: PgStore,
-  assetIdentifier: string
-): Promise<DbFungibleTokenMetadata | undefined> {
-  if (!isFtMetadataEnabled()) {
-    return;
-  }
-  const tokenContractId = assetIdentifier.split('::')[0];
-  const ftMetadata = await db.getFtMetadata(tokenContractId);
-  if (!ftMetadata.found) {
-    if (tokenMetadataErrorMode() === TokenMetadataErrorMode.warning) {
-      logger.warn(`FT metadata not found for token: ${assetIdentifier}`);
-    } else {
-      // TODO: Check if the metadata wasn't found because the contract ABI is not SIP-010
-      // compliant or because there was a recoverable error that prevented the metadata
-      // from being processed.
-      throw new Error(`FT metadata not found for token: ${assetIdentifier}`);
-    }
-  }
-  return ftMetadata.result;
 }
 
 export function getSigners(transaction: StacksTransaction): RosettaAccountIdentifier[] | undefined {
