--- src/tests-tokens-strict/setup.ts
@@ -1,8 +1,0 @@
-import { defaultSetupInit } from '../test-utils/shared-setup';
-
-// ts-unused-exports:disable-next-line
-export default async () => {
-  console.log('Jest - setup..');
-  await defaultSetupInit();
-  console.log('Jest - setup done');
-};
